Add ability to send message to HOST frame from st.user.login()
