+++ frontend/lib/src/hostComm/types.ts
@@ -163,6 +163,10 @@ export type IGuestToHostMessage =
       scriptRunState: ScriptRunState
     }
   | {
+      type: "REDIRECT_TO_URL"
+      url: string
+    }
+  | {
       type: "CUSTOM_PARENT_MESSAGE"
       message: string
     }
